Extract out notification-related code
* Create new NotificationDAL and NotificationController classes, and
move notification-related code from UserDAL and TaskDAL into
NotificationDAL, and from UserController into NotificationController

* Move NotificationSubscriptions from User into UserNotification

* Modify TaskDAL.addComment to accept a Task instead of a task id, and
verify that the target task exists in TaskController when adding a new
comment
