+++ app/org/maproulette/models/dal/TaskDAL.scala
@@ -48,6 +48,7 @@ class TaskDAL @Inject()(override val db: Database,
                         userDAL: Provider[UserDAL],
                         projectDAL: Provider[ProjectDAL],
                         challengeDAL: Provider[ChallengeDAL],
+                        notificationDAL: Provider[NotificationDAL],
                         actions: ActionManager,
                         statusActions: StatusActionManager,
                         ws: WSClient)
@@ -638,7 +639,7 @@ class TaskDAL @Inject()(override val db: Database,
       }
 
       val comment = commentContent.nonEmpty match {
-        case true => Some(addComment(user, task.id, commentContent, actionId))
+        case true => Some(addComment(user, task, commentContent, actionId))
         case false => None
       }
 
@@ -649,7 +650,7 @@ class TaskDAL @Inject()(override val db: Database,
                           (task_id, requested_by, reviewed_by, review_status, reviewed_at)
               VALUES (${task.id}, ${user.id}, ${task.reviewedBy},
                       ${reviewStatus}, ${now})""".executeUpdate()
-        this.generateReviewNotification(user, task.reviewedBy.getOrElse(-1), reviewStatus, task, comment)
+        this.notificationDAL.get().createReviewNotification(user, task.reviewedBy.getOrElse(-1), reviewStatus, task, comment)
       }
       else {
         // Let's note in the task_review_history table that this task was reviewed
@@ -657,7 +658,7 @@ class TaskDAL @Inject()(override val db: Database,
                           (task_id, requested_by, reviewed_by, review_status, reviewed_at)
               VALUES (${task.id}, ${task.reviewRequestedBy}, ${user.id},
                       ${reviewStatus}, ${now})""".executeUpdate()
-        this.generateReviewNotification(user, task.reviewRequestedBy.getOrElse(-1), reviewStatus, task, comment)
+        this.notificationDAL.get().createReviewNotification(user, task.reviewRequestedBy.getOrElse(-1), reviewStatus, task, comment)
       }
 
       this.cacheManager.withOptionCaching { () => Some(task.copy(reviewStatus = Some(reviewStatus))) }
@@ -1565,12 +1566,12 @@ class TaskDAL @Inject()(override val db: Database,
     * Add comment to a task
     *
     * @param user     The user adding the comment
-    * @param taskId   The task that you are adding the comment too
+    * @param task     The task that you are adding the comment too
     * @param comment  The actual comment
     * @param actionId the id for the action if any action associated
     * @param c
     */
-  def addComment(user: User, taskId: Long, comment: String, actionId: Option[Long])(implicit c: Option[Connection] = None): Comment = {
+  def addComment(user: User, task: Task, comment: String, actionId: Option[Long])(implicit c: Option[Connection] = None): Comment = {
     withMRConnection { implicit c =>
       if (StringUtils.isEmpty(comment)) {
         throw new InvalidException("Invalid empty string supplied.")
@@ -1581,13 +1582,13 @@ class TaskDAL @Inject()(override val db: Database,
            |VALUES ({osm_id}, {task_id}, {comment}, {action_id}) RETURNING id, project_id, challenge_id
          """.stripMargin
       SQL(query).on('osm_id -> user.osmProfile.id,
-        'task_id -> taskId,
+        'task_id -> task.id,
         'comment -> comment,
         'action_id -> actionId).as((long("id") ~ long("project_id") ~ long("challenge_id")).*).headOption match {
         case Some(ids) =>
           val newComment =
-            Comment(ids._1._1, user.osmProfile.id, user.name, taskId, ids._1._2, ids._2, DateTime.now(), comment, actionId)
-          this.generateMentionNotifications(user, newComment, taskId)
+            Comment(ids._1._1, user.osmProfile.id, user.name, task.id, ids._1._2, ids._2, DateTime.now(), comment, actionId)
+          this.notificationDAL.get().createMentionNotifications(user, newComment, task)
           newComment
         case None => throw new Exception("Failed to add comment")
       }
@@ -1659,58 +1660,7 @@ class TaskDAL @Inject()(override val db: Database,
     }
   }
 
-  def generateMentionNotifications(fromUser:User, comment:Comment, taskId:Long)(implicit c:Connection=null) = {
-    this.retrieveById(taskId) match {
-      case Some(task) =>
-        // match [@username] (username may contain spaces) or @username (no spaces allowed)
-        val mentionRegex = """\[@([^\]]+)\]|@([\w\d_-]+)""".r.unanchored
 
-        for (m <- mentionRegex.findAllMatchIn(comment.comment)) {
-          // use first non-null group
-          val username = m.subgroups.filter(_ != null).head
-
-          // Retrieve and notify mentioned user
-          userDAL.get().retrieveByOSMUsername(username, User.superUser) match {
-            case Some(mentionedUser) =>
-              userDAL.get().addNotification(mentionedUser.id, User.superUser, UserNotification(
-                -1,
-                userId=mentionedUser.id,
-                notificationType=UserNotification.NOTIFICATION_TYPE_MENTION,
-                fromUsername=Some(fromUser.osmProfile.displayName),
-                taskId=Some(taskId),
-                challengeId=Some(task.parent),
-                targetId=Some(comment.id),
-                extra=Some(comment.comment),
-              ))
-            case None => None
-          }
-        }
-      case None => throw new NotFoundException(s"No task [$taskId] found")
-    }
-  }
-
-  def generateReviewNotification(user: User, forUserId: Int, reviewStatus: Int, task: Task, comment: Option[Comment])(implicit c:Connection=null) = {
-    val notificationType = reviewStatus match {
-      case Task.REVIEW_STATUS_REQUESTED => UserNotification.NOTIFICATION_TYPE_REVIEW_AGAIN
-      case Task.REVIEW_STATUS_APPROVED => UserNotification.NOTIFICATION_TYPE_REVIEW_APPROVED
-      case Task.REVIEW_STATUS_ASSISTED => UserNotification.NOTIFICATION_TYPE_REVIEW_APPROVED
-      case Task.REVIEW_STATUS_REJECTED => UserNotification.NOTIFICATION_TYPE_REVIEW_REJECTED
-    }
-
-    userDAL.get().addNotification(forUserId, User.superUser, UserNotification(
-      -1,
-      userId=forUserId,
-      notificationType=notificationType,
-      fromUsername=Some(user.osmProfile.displayName),
-      description=Some(reviewStatus.toString()),
-      taskId=Some(task.id),
-      challengeId=Some(task.parent),
-      extra=comment match {
-        case Some(c) => Some(c.comment)
-        case None => None
-      }
-    ))
-  }
 
   private def getTaskGeometries(id: Long)(implicit c: Option[Connection] = None): String = this.taskGeometries(id, "task_geometries")
