+++ app/org/maproulette/session/dal/UserDAL.scala
@@ -15,9 +15,9 @@ import org.joda.time.DateTime
 import org.maproulette.Config
 import org.maproulette.cache.CacheManager
 import org.maproulette.data.UserType
-import org.maproulette.exception.{NotFoundException, InvalidException}
+import org.maproulette.exception.NotFoundException
 import org.maproulette.models.dal.{BaseDAL, ChallengeDAL, ProjectDAL, TaskDAL}
-import org.maproulette.models.{Challenge, Project, Task, UserNotification, UserNotificationEmail}
+import org.maproulette.models.{Challenge, Project, Task}
 import org.maproulette.permissions.Permission
 import org.maproulette.session._
 import org.maproulette.utils.{Crypto, Utils}
@@ -120,53 +120,6 @@ class UserDAL @Inject()(override val db: Database,
       get[List[Int]]("group_types") map {
       case projectId ~ osmId ~ displayName ~ avatarURL ~ groupTypes =>
         ProjectManager(projectId, osmId, displayName, avatarURL.getOrElse(""), groupTypes)
-    }
-  }
-
-  // The anorm row parser for user notifications
-  val userNotificationParser: RowParser[UserNotification] = {
-    get[Long]("user_notifications.id") ~
-    get[Long]("user_notifications.user_id") ~
-    get[Int]("user_notifications.notification_type") ~
-    get[DateTime]("user_notifications.created") ~
-    get[DateTime]("user_notifications.modified") ~
-    get[Option[String]]("user_notifications.description") ~
-    get[Option[String]]("user_notifications.from_username") ~
-    get[Option[String]]("challenges.name") ~
-    get[Boolean]("user_notifications.is_read") ~
-    get[Int]("user_notifications.email_status") ~
-    get[Option[Long]]("user_notifications.task_id") ~
-    get[Option[Long]]("user_notifications.challenge_id") ~
-    get[Option[Long]]("user_notifications.project_id") ~
-    get[Option[Long]]("user_notifications.target_id") ~
-    get[Option[String]]("user_notifications.extra") map {
-      case id ~ userId ~ notificationType ~ created ~ modified ~ description ~ fromUsername ~ challengeName ~ isRead ~ emailStatus ~ taskId ~ challengeId ~ projectId ~ targetId ~ extra =>
-        new UserNotification(id, userId, notificationType, created, modified, description, fromUsername, challengeName, isRead, emailStatus, taskId, challengeId, projectId, targetId, extra)
-    }
-  }
-
-  val userNotificationEmailParser: RowParser[UserNotificationEmail] = {
-    get[Long]("user_notifications.id") ~
-    get[Long]("user_notifications.user_id") ~
-    get[Int]("user_notifications.notification_type") ~
-    get[DateTime]("user_notifications.created") ~
-    get[Int]("user_notifications.email_status") map {
-      case id ~ userId ~ notificationType ~ created ~ emailStatus =>
-        new UserNotificationEmail(id, userId, notificationType, created, emailStatus)
-    }
-  }
-
-  // The anorm row parser for user's subscriptions to notifications
-  val notificationSubscriptionParser: RowParser[NotificationSubscriptions] = {
-    get[Long]("id") ~
-    get[Long]("user_id") ~
-    get[Int]("system") ~
-    get[Int]("mention") ~
-    get[Int]("review_approved") ~
-    get[Int]("review_rejected") ~
-    get[Int]("review_again") map {
-      case id ~ userId ~ system ~ mention ~ reviewApproved ~ reviewRejected ~ reviewAgain =>
-        NotificationSubscriptions(id, userId, system, mention, reviewApproved, reviewRejected, reviewAgain),
     }
   }
 
@@ -1004,192 +957,6 @@ class UserDAL @Inject()(override val db: Database,
     this.projectDAL.retrieveByName(homeName) match {
       case Some(project) => project
       case None => throw new NotFoundException("You should never get this exception, Home project should always exist for user.")
-    }
-  }
-
-  /**
-   * Retrieves the user's notification subscriptions
-   *
-   * @param userId The id of the subscribing user
-   * @param user The user making the request
-   * @return
-   */
-  def getNotificationSubscriptions(userId:Long, user:User)(implicit c:Connection=null): NotificationSubscriptions = {
-    this.permission.hasReadAccess(UserType(), user)(userId)
-    withMRConnection { implicit c =>
-      SQL(
-        s"""SELECT * FROM user_notification_subscriptions
-            WHERE user_id=${userId} LIMIT 1"""
-      ).as(notificationSubscriptionParser.*).headOption match {
-        case Some(subscription) => subscription
-        case None =>
-          // Default to subscribing to all notifications, but with no emails
-          NotificationSubscriptions(-1, userId, UserNotification.NOTIFICATION_EMAIL_NONE,
-                                                UserNotification.NOTIFICATION_EMAIL_NONE,
-                                                UserNotification.NOTIFICATION_EMAIL_NONE,
-                                                UserNotification.NOTIFICATION_EMAIL_NONE,
-                                                UserNotification.NOTIFICATION_EMAIL_NONE)
-      }
-    }
-  }
-
-  /**
-   * Updates the user's notification subscriptions
-   *
-   * @param userId The id of the subscribing user
-   * @param user The user making the request
-   * @param subscriptions The updated subscriptions
-   * @return
-   */
-  def updateNotificationSubscriptions(userId: Long, user: User, subscriptions: NotificationSubscriptions)(implicit c:Connection=null) = {
-    this.permission.hasWriteAccess(UserType(), user)(userId)
-    withMRTransaction { implicit c =>
-      // Upsert new subscription settings
-      SQL(
-        s"""INSERT INTO user_notification_subscriptions (user_id, system, mention, review_approved, review_rejected, review_again)
-            VALUES({userId}, {system}, {mention}, {reviewApproved}, {reviewRejected}, {reviewAgain})
-            ON CONFLICT (user_id) DO
-            UPDATE SET system=EXCLUDED.system, mention=EXCLUDED.mention, review_approved=EXCLUDED.review_approved, review_rejected=EXCLUDED.review_rejected, review_again=EXCLUDED.review_again"""
-      ).on(
-        'userId -> userId,
-        'system -> subscriptions.system,
-        'mention -> subscriptions.mention,
-        'reviewApproved -> subscriptions.reviewApproved,
-        'reviewRejected -> subscriptions.reviewRejected,
-        'reviewAgain -> subscriptions.reviewAgain,
-      ).executeUpdate()
-    }
-  }
-
-  /**
-   * Adds a notification for the given user, updating its email settings based
-   * on the user's current email preferences. If the user is not subscribed to
-   * the type of notification given then it is simply ignored.
-   *
-   * @param userId The id of the user to notify
-   * @param user The user making the request
-   * @param notification The notification to add
-   * @return
-   */
-  def addNotification(userId:Long, user:User, notification:UserNotification)(implicit c:Connection=null) = {
-    this.permission.hasWriteAccess(UserType(), user)(userId)
-    val subscriptions = this.getNotificationSubscriptions(userId, user)
-    val subscriptionType = notification.notificationType match {
-      case UserNotification.NOTIFICATION_TYPE_SYSTEM => subscriptions.system
-      case UserNotification.NOTIFICATION_TYPE_MENTION => subscriptions.mention
-      case UserNotification.NOTIFICATION_TYPE_REVIEW_APPROVED => subscriptions.reviewApproved
-      case UserNotification.NOTIFICATION_TYPE_REVIEW_REJECTED => subscriptions.reviewRejected
-      case UserNotification.NOTIFICATION_TYPE_REVIEW_AGAIN => subscriptions.reviewAgain
-      case _ => throw new InvalidException("Invalid notification type")
-    }
-
-    // Guard against ignored notification type
-    subscriptionType match {
-      case UserNotification.NOTIFICATION_IGNORE => None // nothing to do
-      case _ =>
-        notification.userId = userId
-        notification.emailStatus = subscriptionType
-        notification.isRead = false
-        this.withMRTransaction { implicit c =>
-          val query =
-            """INSERT INTO user_notifications (user_id, notification_type, description, from_username, is_read,
-                                               email_status, task_id, challenge_id, project_id, target_id, extra)
-               VALUES ({user_id}, {notification_type}, {description}, {from_username}, {is_read},
-                       {email_status}, {task_id}, {challenge_id}, {project_id}, {target_id}, {extra})
-               RETURNING *"""
-
-          val newNotification = SQL(query).on(
-            'user_id -> notification.userId,
-            'notification_type -> notification.notificationType,
-            'description -> notification.description,
-            'from_username -> notification.fromUsername,
-            'is_read -> false,
-            'email_status -> notification.emailStatus,
-            'task_id -> notification.taskId,
-            'challenge_id -> notification.challengeId,
-            'project_id -> notification.projectId,
-            'target_id -> notification.targetId,
-            'extra -> notification.extra
-          ).execute()
-        }
-    }
-  }
-
-  /**
-   * Marks as read the given notifications owned by the given userId
-   *
-   * @param userId The id of the user that owns the notifications
-   * @param user The user making the request
-   * @param notificationIds The ids of the notifications to be marked read
-   */
-  def markNotificationsRead(userId: Long, user: User, notificationIds: List[Long])(implicit c:Connection=null) = {
-    this.permission.hasWriteAccess(UserType(), user)(userId)
-      this.withMRTransaction { implicit c =>
-        val query =
-          s"""UPDATE user_notifications SET is_read=true
-              WHERE user_notifications.user_id={user_id}
-              ${this.getLongListFilter(Some(notificationIds), "user_notifications.id")}"""
-        SQL(query).on('user_id -> userId).execute()
-      }
-  }
-
-  /**
-   * Deletes the given notifications owned by the given userId
-   *
-   * @param userId The id of the user that owns the notifications
-   * @param user The user making the request
-   * @param notificationIds The ids of the notifications to delete
-   * @return
-   */
-  def deleteNotifications(userId: Long, user: User, notificationIds: List[Long])(implicit c:Connection=null) = {
-    this.permission.hasWriteAccess(UserType(), user)(userId)
-      this.withMRTransaction { implicit c =>
-        val query =
-          s"""DELETE from user_notifications
-              WHERE user_notifications.user_id={user_id}
-              ${this.getLongListFilter(Some(notificationIds), "user_notifications.id")}"""
-        SQL(query).on('user_id -> userId).execute()
-      }
-  }
-
-  /**
-   * Retrieves the user notifications sent to the given userId
-   */
-  def getUserNotifications(userId:Long, user:User, limit:Int=Config.DEFAULT_LIST_SIZE, offset:Int=0,
-                           orderColumn:String="is_read", orderDirection:String="ASC",
-                           notificationType:Option[Int]=None, isRead:Option[Boolean]=None,
-                           fromUsername:Option[String]=None, challengeId:Option[Long]=None)
-                          (implicit c:Connection=null) : List[UserNotification] = {
-    this.permission.hasReadAccess(UserType(), user)(userId)
-    withMRConnection { implicit c =>
-      val whereClause = new StringBuilder("WHERE user_id = {userId}")
-      appendInWhereClause(whereClause, getOptionalFilter(notificationType, "notification_type", "notificationType"))
-      appendInWhereClause(whereClause, getOptionalFilter(isRead, "is_read", "isRead"))
-      appendInWhereClause(whereClause, getOptionalFilter(challengeId, "challenge_id", "challengeId"))
-      appendInWhereClause(whereClause, getOptionalMatchFilter(fromUsername, "from_username", "fromUsername"))
-
-      // In addition to the requested sort, we always add a sort by created desc
-      // (unless created was the requested sort column)
-      var orderClause = this.order(Some(orderColumn), orderDirection)
-      if (orderColumn != "created") {
-        orderClause ++= ", created desc"
-      }
-
-      val query = s"""
-         |SELECT user_notifications.*, challenges.name
-         |FROM user_notifications
-         |LEFT OUTER JOIN challenges on user_notifications.challenge_id = challenges.id
-         |${whereClause}
-         |${orderClause}
-         |LIMIT ${sqlLimit(limit)} OFFSET $offset
-       """.stripMargin
-       SQL(query).on(
-         'userId -> userId,
-         'notificationType -> notificationType,
-         'isRead -> isRead,
-         'challengeId -> challengeId,
-         'fromUsername -> fromUsername,
-       ).as(userNotificationParser.*)
     }
   }
 }
